API_SIMIT_Mail/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-cors-headers==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
lxml==4.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.2
psycopg2-binary==2.8.2
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
spyne==2.13.2a0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.2
